+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Dev
+- New utils for React
+
 # Version 1.0.1
 - Fixed type conflicts and function declarations in `lib.dom.d.ts`